+++ objectbox/CHANGELOG.md
@@ -40,6 +40,14 @@
   int get computedValue => TODO;
   set computedValue(int value) => TODO;
   ```
+* Support creating file-less in-memory databases, for example for testing. To create one pass an 
+  in-memory identifier together with `Store.inMemoryPrefix` as the `directory`:
+  ```dart
+   final inMemoryStore =
+       Store(getObjectBoxModel(), directory: "${Store.inMemoryPrefix}test-db");
+  ```
+  See the `Store` documentation for details.
+* Add `Store.removeDbFiles()` to conveniently delete database files or an in-memory database.
 
 ## 2.4.0 (2023-12-13)